refactor shared stuff into inkstitch.py
