+++ embroider_params.py
@@ -13,7 +13,8 @@ import wx
 from wx.lib.scrolledpanel import ScrolledPanel
 from collections import defaultdict
 import inkex
-from embroider import Param, EmbroideryElement, Fill, AutoFill, Stroke, SatinColumn, descendants
+from inkstitch import Param, EmbroideryElement, get_nodes
+from embroider import Fill, AutoFill, Stroke, SatinColumn
 from functools import partial
 from itertools import groupby
 from embroider_simulate import EmbroiderySimulator
@@ -623,17 +624,6 @@ class EmbroiderParams(inkex.Effect):
         self.cancelled = False
         inkex.Effect.__init__(self, *args, **kwargs)
 
-    def get_nodes(self):
-        if self.selected:
-            nodes = []
-            for node in self.document.getroot().iter():
-                if node.get("id") in self.selected:
-                    nodes.extend(descendants(node))
-        else:
-            nodes = descendants(self.document.getroot())
-
-        return nodes
-
     def embroidery_classes(self, node):
         element = EmbroideryElement(node)
         classes = []
@@ -651,10 +641,10 @@ class EmbroiderParams(inkex.Effect):
         return classes
 
     def get_nodes_by_class(self):
-        nodes = self.get_nodes()
+        nodes = get_nodes(self)
         nodes_by_class = defaultdict(list)
 
-        for z, node in enumerate(self.get_nodes()):
+        for z, node in enumerate(nodes):
             for cls in self.embroidery_classes(node):
                 element = cls(node)
                 element.order = z
